move save_snapshot to shared
