+++ conoha_client/_shared/__init__.py
@@ -1,1 +1,3 @@
 """featuresに依存した共有ライブラリ."""
+from .renforced_vm import *  # noqa: F403
+from .snapshot import *  # noqa: F403
