+++ conoha_client/_shared/renforced_vm/__init__.py
@@ -1,1 +1,3 @@
 from .cli import list_vm_cli, reinforced_vm_cli  # noqa: F401
+from .domain import ReinforcedVM  # noqa: F401
+from .query import find_reinforced_vm_by_id  # noqa: F401
